refactor SummaryPane into show-results
