+++ src/views/show-results/SummaryPane.vue
